decomposition of slices
