+++ src/components/location-list/location-list.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { LOCATIONS, DEFAULT_CITY } from '../../const';
 import { CityName } from '../../types';
-import { changeCity, selectActiveCity } from '../../store/slices/offers';
+import { changeCity, selectActiveCity } from '../../store/slices/city';
 import { useAppDispatch, useAppSelector } from '../../store/hooks';
 import LocationItem from './location-item';
